Fix: all pre-submission critical and moderate issues
Critical:
- config.py: fix stale comment (bge-large -> bge-small)
- structural.py: hashlib.md5 replaces hash() for PYTHONHASHSEED determinism
- structural.py: word-boundary regex fixes short-token false positives (rag/storage, e5/e5500)
- structural.py: remove vector DBs from _PRE_LLM_TERMS (Pinecone != pre-LLM production ML)
- structural.py: hoist nlp/langchain/pre_llm tuples to module scope (100K alloc removed)
- structural.py: local RESEARCH_TITLE_TERMS shadow -> module-level _RESEARCH_FLAVORED_TITLES
- loading.py: add 6-month duration filter (README claim now matches code)
- embed.py: float16 -> float32 (NDCG@10 tiebreak precision restored)
- README.md: model name MiniLM->BAAI/bge-small-en-v1.5
- README.md: weight table corrected (education_tier 0.05, logistics 0.05)

Moderate:
- pipeline.py: _InvertedStr hack -> (-score, cid, scored) explicit heap tuples
- TODO_MANUAL_STEPS.md: fix MiniLM ref, education claim, artifact size
- .gitignore: add dump files, update artifact size comment
- Delete PROJECT_WORKING_EXPLANATION.md (wrong model, made-up timing)

All 26 tests pass.

+++ embed.py
@@ -78,8 +78,8 @@ def main() -> None:
             print(f"  {n:>6} candidates embedded  ({elapsed/60:.1f} min, {n/elapsed:.0f}/s)")
     flush()
 
-    matrix = np.vstack(chunks) if chunks else np.zeros((0, config.EMBEDDING_DIM), np.float16)
-    np.save(out_dir / "candidate_embeddings.npy", matrix.astype(np.float16))
+    matrix = np.vstack(chunks) if chunks else np.zeros((0, config.EMBEDDING_DIM), np.float32)
+    np.save(out_dir / "candidate_embeddings.npy", matrix.astype(np.float32))
     (out_dir / "candidate_ids.json").write_text(json.dumps(ids), encoding="utf-8")
 
     print(f"Done: {len(ids)} embeddings -> {out_dir}/ in {(time.time()-started)/60:.1f} min")
